Move the handling of 'interactive' packages to PM layer
It makes more sense as it already takes care of the list of minimal
packages. In practice the only difference is that it will now use the host
system urpmi to install it, instead of running it from the chroot. (ie. it
allows using local medias for installing these packages)