jurt-build now stops when some of the builds fail
The main reason to build many source packages in one invocation is to allow
packages to be built using the binaries provided by the previous ones, so
any failed build may potentially result in packages with incorrect build
dependencies.

Added an option --keep-building to force jurt to keep building in case any
package fails to build.